+++ bigtable/docs/specs/SESSION_POOL_SPEC.md
@@ -81,7 +81,7 @@ Any new pool method that reads `p.picker.Name()` from a method called by `Checko
 **`TableShim` (`bigtable/table_shim.go`) — mechanism layer.**
 - Implements the public `TableAPI` — this is what user code holds when running mixed-mode. Owns `(classic TableAPI, session SessionTableApi, diverter *Diverter)`. Any of `session` / `diverter` may be nil → **shim degrades to classic-only** silently. This is the fallback contract when session support is not enabled or the pool failed to open.
 - Per-call routing rule: `if !t.useSession() { classic } else { session }`. `useSession()` = `session != nil && diverter != nil && diverter.UseSession()`.
-- **`session` is a `*sessionTableHandle` that self-heals across cache eviction.** TableShim caches the handle pointer at Open time and holds it for the shim's lifetime. When the `sessionTableCache` TTL sweeper evicts that handle (or any owner calls `handle.Close()`), the handle's `evicted atomic.Bool` flips to true BEFORE its underlying api is torn down; the next `handle.ReadRow`/`MutateRow` observes the flag on its atomic-Load fast path and routes through `cache.getOrOpen` to fetch (or mint) a live successor, transparently dispatching the RPC there. TableShim never sees the eviction — its cached pointer stays valid forever, and per-RPC cost stays at one atomic Load on the happy path. See `session_table_cache.go` for the invariants; guarded by `TestSessionTableHandle_EvictedSelfHeals`.
+- **`session` is a `*sessionTableHandle` that self-heals across cache eviction.** The pointer TableShim caches at Open time stays valid for the shim's lifetime; the wrapper routes evicted RPCs through `cache.getOrOpen`. Go-specific hazard — Java's session-table cache has no TTL sweeper, so this handle-past-eviction gap doesn't exist there and Design C has no direct Java analog. Mechanism lives in `session_table_cache.go` (see `sessionTableHandle.dispatch` + `resolveSuccessor`); guarded by `TestSessionTableHandle_EvictedSelfHeals` / `TestSessionTableHandle_SweeperEvictionSelfHeals`.
 - Owns **all proto ↔ `bigtable.Row` conversion** at the boundary — the `internal/session` package stays proto-native (never sees `bigtable.Row`, `Mutation`, `Filter`, etc.). This is how the two data planes stay decoupled.
 - Method routability is **fixed by shape**, not by config — the shim MUST NOT attempt to route an operation whose vRPC equivalent doesn't exist. Enforced today as:
